+++ eventkit_cloud/utils/helpers.py
@@ -1,8 +1,7 @@
 from contextlib import contextmanager
-from functools import wraps
+
 import logging
 import os
-import time
 
 from django.conf import settings
 from django.core.cache import cache
@@ -10,51 +9,6 @@ from django.core.cache import cache
 from eventkit_cloud.utils.mapproxy import mapproxy_config_keys_index
 
 logger = logging.getLogger()
-
-MAX_DB_CONNECTION_RETRIES = 8
-TIME_DELAY_BASE = 2  # Used for exponential delays (i.e. 5^y) at 8 would be about 4 minutes 15 seconds max delay.
-
-# The retry here is an attempt to mitigate any possible dropped connections. We chose to do a limited number of
-# retries as retrying forever would cause the job to never finish in the event that the database is down. An
-# improved method would perhaps be to see if there are connection options to create a more reliable connection.
-# We have used this solution for now as I could not find options supporting this in the gdal documentation.
-
-
-def retry(f):
-    @wraps(f)
-    def wrapper(*args, **kwds):
-
-        attempts = MAX_DB_CONNECTION_RETRIES
-        exc = None
-        while attempts:
-            try:
-                return_value = f(*args, **kwds)
-                if not return_value:
-                    logger.error("The function {0} failed to return any values.".format(getattr(f, "__name__")))
-                    raise Exception("The process failed to return any data, please contact an administrator.")
-                return return_value
-            except Exception as e:
-                logger.error("The function {0} threw an error.".format(getattr(f, "__name__")))
-                logger.error(str(e))
-                exc = e
-
-                if getattr(settings, "TESTING", False):
-                    # Don't wait/retry when running tests.
-                    break
-                attempts -= 1
-                logger.info(e)
-                if "canceled" in str(e).lower():
-                    # If task was canceled (as opposed to fail) don't retry.
-                    logger.info("The task was canceled ")
-                    attempts = 0
-                else:
-                    if attempts:
-                        delay = TIME_DELAY_BASE ** (MAX_DB_CONNECTION_RETRIES - attempts + 1)
-                        logger.error(f"Retrying {str(attempts)} more times, sleeping for {delay}...")
-                        time.sleep(delay)
-        raise exc
-
-    return wrapper
 
 
 def get_relative_path_from_staging(staging_path):
